requirements.txt
tensorboard
torchvision
tqdm
matplotlib
pandas
scipy
numpy
seaborn
torch
Pillow
scikit_learn
polyaxon-client
opencv-python
